src/requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.16
langchain-huggingface==0.1.2
chromadb==0.6.3
streamlit==1.45.0
python-dotenv==1.1.0
sentence-transformers==4.1.0
tiktoken==0.9.0
openai==1.77.0
unstructured==0.17.2
pypdf==5.4.0
watchdog==6.0.0